Update saconvolution2d.py
A fix so that the layer honours the arguments passed to the constructor in the build process even if that itself is not called from the constructor - i.e. the input_shape is not yet known. Other arguments such as padding, regularisation et.c. should still be honoured while building unless for some reason overridden in the call to build. In the latter case the arguments for the constructor and the build method are merged before being passed to the Conv2D constructor.