Update saconvolution2d.py
Added function (not method or class method or anything, just a plain vanilla function):

 inject_saconv2d(model, layer_list)

where model is a preexisting model and layer_list is a list of Conv2D layers that should be transparently (hopefully) replaced by SAConvolution2D. It should keep its configuration and weights (and even name) but now be "spatially aware" and have some newly initialised weights for incorporating this in the convolution. Obviously these weights will have to be trained if there should be any point in it but all other weights start from where they were (including the old weights from the Conv2D layer).

The layers to replace (which have to be Conv2D) can be listed by number, name, regex (one entry can match several layers) or any combination of these.